separate out plotting
